Broke up gp.py and changed some getter/setters to properties
