+++ clay/core/__init__.py
@@ -3,4 +3,4 @@ from .gp import gp
 from .log import logger
 from .settings import settings_manager
 from .osd import osd_manager
-from .mpris2 import MPRIS2
+from .mpris2 import mpris2_manager
