+++ clay/core/gp/__init__.py
@@ -1,0 +1,20 @@
+# This file is part of Clay.
+# Copyright (C) 2018, Andrew Dunbai & Clay Contributors
+#
+# Clay is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# Clay is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with Clay. If not, see <https://www.gnu.org/licenses/>.
+"""
+This file contains the module organization of the Google Play specific stuff
+"""
+
+from .client import gp
