+++ clay/ui/urwid/playbar.py
@@ -117,7 +117,7 @@ class PlayBar(urwid.Pile):
         """
         Return the style for current playback state.
         """
-        if player.is_loading or player.is_playing:
+        if player.loading or player.playing:
             return 'title-playing'
         return 'title-idle'
 
@@ -136,8 +136,8 @@ class PlayBar(urwid.Pile):
         return (self.get_style(), u' {} {} - {} {} [{:02d}:{:02d} / {:02d}:{:02d}]'.format(
             # u'|>' if player.is_playing else u'||',
             # self.get_rotating_bar(),
-            u'\u2505' if player.is_loading
-            else u'\u25B6' if player.is_playing
+            u'\u2505' if player.loading
+            else u'\u25B6' if player.playing
             else u'\u25A0',
             track.artist,
             track.title,
@@ -158,14 +158,14 @@ class PlayBar(urwid.Pile):
         self.progressbar.set_progress(player.play_progress)
         self.progressbar.set_done_style(
             'progressbar_done'
-            if player.is_playing
+            if player.playing
             else 'progressbar_done_paused'
         )
         self.shuffle_el.attr = 'flag-active' \
-            if player.get_is_random() \
+            if player.random \
             else 'flag'
         self.repeat_el.attr = 'flag-active' \
-            if player.get_is_repeat_one() \
+            if player.repeat_one \
             else 'flag'
         self.app.redraw()
