+++ tox.ini
@@ -1,6 +1,5 @@
 [tox]
 envlist =
-    py27
     py36
 
 [testenv]
